+++ ec2-spot-history-notebook/cdk/app.py
@@ -1,77 +1,10 @@
 #!/usr/bin/env python3
 
-"""
 from aws_cdk import core
-from cdk.cdk_stack import CdkStack
-app = core.App()
-CdkStack(app, "cdk")
-app.synth()
-"""
 
-import os
-import boto3 
-import uuid
-
-DEFAULT_GITHUB_REPO="https://github.com/ruecarlo/ec2-spot-labs.git"
-
-profile = os.getenv("AWS_PROFILE", "default")
-session = boto3.Session(profile_name=profile)
-my_region = session.region_name
-my_acc_id = session.client('sts').get_caller_identity().get('Account')
-ec2client = session.client('ec2')
-smclient = session.client('sagemaker')
-
-#USER VARIABLES
-default_vpc = [x['VpcId'] for x in ec2client.describe_vpcs()['Vpcs'] if x['IsDefault']][0]
-default_sg = [x['GroupId'] for x in ec2client.describe_security_groups(Filters=[{'Name':'vpc-id','Values':[default_vpc,]},])['SecurityGroups'] if x['GroupName']=='default']
-default_subnet = [x['SubnetId'] for x in ec2client.describe_subnets(Filters=[{'Name':'vpc-id','Values':[default_vpc,]},])['Subnets'] if x['DefaultForAz']][0]
-
-
-from aws_cdk import (
-    aws_iam as iam_,
-    aws_sagemaker as sagemaker_,
-    core)
-
-class SpotHistoryNotebook(core.Stack):
-
-    def __init__(self, app: core.App, id: str) -> None:
-        super().__init__(app, id)
-
-        # Create role for Notebook instance
-        nrole = iam_.Role(
-            self,
-            "notebookAccessRole"
-        )
-
-        nrole.add_managed_policy(iam_.ManagedPolicy.from_aws_managed_policy_name('AmazonEC2ReadOnlyAccess'))
-
-
-        notebook_uuid=str(uuid.uuid4())
-        notebook_uuid=str(notebook_uuid[0:notebook_uuid.find('-')])
-        notebook_instance_id = 'spot-history-notebook-'+notebook_uuid
-        print("id {}".format(notebook_instance_id))
-
-        notebook_instance = sagemaker_.CfnNotebookInstance(
-            self,
-            notebook_instance_id,
-            instance_type='ml.m5.large',
-            volume_size_in_gb=10,
-            security_group_ids=default_sg,
-            subnet_id=default_subnet,
-            notebook_instance_name=notebook_instance_id,
-            role_arn=nrole.role_arn,
-            default_code_repository=DEFAULT_GITHUB_REPO
-        )
-
-        core.CfnOutput(
-            self,
-            "Notebook Name",
-            value=notebook_instance.notebook_instance_name,
-            description="Notebook Instance Name",  # Optional
-        )
-
+from spot_historic_notebook.spot_historic_notebook_stack import CdkSpotHistoricNotebookStack
 
 
 app = core.App()
-SpotHistoryNotebook(app, "SpotHistoryNotebook")
+CdkSpotHistoricNotebookStack(app, "spot_historic_notebook")
 app.synth()
