--- ec2-spot-history-notebook/cdk/cdk/cdk.egg-info/PKG-INFO
@@ -1,80 +1,0 @@
-Metadata-Version: 2.1
-Name: cdk
-Version: 0.0.1
-Summary: An empty CDK Python app
-Home-page: UNKNOWN
-Author: author
-License: UNKNOWN
-Description: 
-        # Welcome to your CDK Python project!
-        
-        This is a blank project for Python development with CDK.
-        
-        The `cdk.json` file tells the CDK Toolkit how to execute your app.
-        
-        This project is set up like a standard Python project.  The initialization
-        process also creates a virtualenv within this project, stored under the .env
-        directory.  To create the virtualenv it assumes that there is a `python3`
-        (or `python` for Windows) executable in your path with access to the `venv`
-        package. If for any reason the automatic creation of the virtualenv fails,
-        you can create the virtualenv manually.
-        
-        To manually create a virtualenv on MacOS and Linux:
-        
-        ```
-        $ python3 -m venv .env
-        ```
-        
-        After the init process completes and the virtualenv is created, you can use the following
-        step to activate your virtualenv.
-        
-        ```
-        $ source .env/bin/activate
-        ```
-        
-        If you are a Windows platform, you would activate the virtualenv like this:
-        
-        ```
-        % .env\Scripts\activate.bat
-        ```
-        
-        Once the virtualenv is activated, you can install the required dependencies.
-        
-        ```
-        $ pip install -r requirements.txt
-        ```
-        
-        At this point you can now synthesize the CloudFormation template for this code.
-        
-        ```
-        $ cdk synth
-        ```
-        
-        To add additional dependencies, for example other CDK libraries, just add
-        them to your `setup.py` file and rerun the `pip install -r requirements.txt`
-        command.
-        
-        ## Useful commands
-        
-         * `cdk ls`          list all stacks in the app
-         * `cdk synth`       emits the synthesized CloudFormation template
-         * `cdk deploy`      deploy this stack to your default AWS account/region
-         * `cdk diff`        compare deployed stack with current state
-         * `cdk docs`        open CDK documentation
-        
-        Enjoy!
-        
-Platform: UNKNOWN
-Classifier: Development Status :: 4 - Beta
-Classifier: Intended Audience :: Developers
-Classifier: License :: OSI Approved :: Apache Software License
-Classifier: Programming Language :: JavaScript
-Classifier: Programming Language :: Python :: 3 :: Only
-Classifier: Programming Language :: Python :: 3.6
-Classifier: Programming Language :: Python :: 3.7
-Classifier: Programming Language :: Python :: 3.8
-Classifier: Topic :: Software Development :: Code Generators
-Classifier: Topic :: Utilities
-Classifier: Typing :: Typed
-Requires-Python: >=3.6
-Description-Content-Type: text/markdown
