--- ec2-spot-history-notebook/cdk/cdk/cdk.egg-info/requires.txt
@@ -1,4 +1,0 @@
-aws-cdk.core
-aws-cdk.aws-iam
-aws-cdk.aws-sagemaker
-boto3
