--- ec2-spot-history-notebook/cdk/cdk/cdk.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-
