--- ec2-spot-history-notebook/cdk/cdk/cdk_stack.py
@@ -1,9 +1,0 @@
-from aws_cdk import core
-
-
-class CdkStack(core.Stack):
-
-    def __init__(self, scope: core.Construct, id: str, **kwargs) -> None:
-        super().__init__(scope, id, **kwargs)
-
-        # The code that defines your stack goes here
